Add rustic windmill and related resources
Rustic Windmill is an entry level "western-style" windmill
that uses a new resource "small stators" to produce energy
instead of "small engines". It produces half the amount of
energy than the modern variant. Small stators can be found
around the world in various electric devices and cars. It
scraps into brass, so it is also an alternative source for
this always rare metal. Can also be crafted in forges.